# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: EventosAPI
host: localhost:10010
...
basePath: /
schemes: [http, https]
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /eventos:
    x-swagger-router-controller: evento.controller
    
    get:
      operationId: getEventos
      parameters:
        - name: tags
          in: query
          required: false
          type: array
          items:
            type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EventoList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      operationId: postEvento
      parameters:
        - name: evento
          in: body
          required: true
          schema:
            $ref: "#/definitions/Evento"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/EventoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /eventos/{id}:
    x-swagger-router-controller: evento.controller
  
    get:
      operationId: getEventoById
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: success
          schema:
            $ref: "#/definitions/EventoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      operationId: putEvento
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: evento
          in: body
          schema:
            $ref: "#/definitions/Evento"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EventoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      operationId: deleteEvento
      parameters:
        - name: id
          in: path 
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  DeleteResponse:
    type: object
    properties:
      success:
        type: number
        description: Returns 1 if successful
      description:
        type: string
        description: A short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
        
# Login
...
  Evento:
    type: object
    properties:
      nombre: 
        type: string
      descripcion:
        type: string
      invitados:
        type: string
      fecha: 
        type: string
      hora:
        type: string
      tags:
        type: array
        items:
          type: string
      foto:
        type: string
      realizado:
        type: boolean
    required:
      - nombre
      - descripcion
      - invitados
      - fecha
      - tags
      - realizado
  EventoResponse:
    type: object
    properties:
      id: 
        type: number
      nombre: 
        type: string
      descripcion:
        type: string
      invitados:
        type: string
      fecha: 
        type: string
      hora:
        type: string
      tags:
        type: array
        items:
          type: string
      foto:
        type: string
      realizado:
        type: boolean
    required:
      - nombre
      - descripcion
      - invitados
      - fecha
      - tags
      - realizado
  EventoList:
    properties:
      eventos:
        type: array
        items:
          $ref: "#/definitions/EventoResponse"  
          
# Usuario-Evento
...